+++ examples/two-step-transfer/src/lib.rs
@@ -62,7 +62,7 @@ struct TransferRequest {
 #[derive(Serialize, SchemaType)]
 struct InitParams {
     /// Who is authorized to withdraw funds from this lockup (must be non-empty)
-    #[concordium(set_size_length = 1)]
+    #[concordium(size_length = 1)]
     account_holders: BTreeSet<AccountAddress>,
 
     /// How many of the account holders need to agree before funds are released
@@ -89,7 +89,7 @@ pub struct State {
     // Requests which have not been dropped due to timing out or due to being agreed to yet
     // The request ID, the associated amount, when it times out, who is making the transfer and
     // which account holders support this transfer
-    #[concordium(map_size_length = 2)]
+    #[concordium(size_length = 2)]
     requests: BTreeMap<TransferRequestId, TransferRequest>,
 }